+++ src/main/java/event/load/LoadEvent.java
@@ -1,4 +1,4 @@
-package event;
+package event.load;
 
 public class LoadEvent<T>{
